split out sidebar child components for easier datapassing
